create firebase.js
・firebaseの設定を別ファイルに転載する。
